# api/openapi.yaml
openapi: 3.0.3
info:
  title: Helling API
  version: 0.2.0
  description: |
    Helling is a self-hosted, single-node container and VM management platform.
    This is the v0.2 Platform Core bridge: authentication, instance CRUD,
    operation polling, schedules, webhooks, host firewall, events, audit, and
    user Incus scope assignment.

    All endpoints under `/v1/` are stable within v0.2. Breaking changes ship under `/v2/`.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  contact:
    name: Bizarre Industries
    url: https://github.com/Bizarre-Industries/helling
servers:
  - url: https://localhost/v1
    description: Local installation behind the Caddy edge service
  - url: http+unix:///run/helling/api.sock/v1
    description: Direct socket access (CLI, dev)
...
security:
  - cookieAuth: []
paths:
...
  /events:
    get:
      operationId: listEvents
      tags:
        - events
      summary: Stream or list events
      description: |
        Streams Server-Sent Events when requested with
        `Accept: text/event-stream`; otherwise returns a JSON snapshot for CLI
        and dashboard hydration.
      parameters:
        - in: query
          name: limit
          description: Maximum number of recent events for JSON snapshot mode.
          schema:
            type: integer
            description: Maximum number of recent events.
            minimum: 1
            maximum: 500
            default: 50
            example: 50
        - in: query
          name: type
          description: Optional exact or prefix event type filter.
          schema:
            type: string
            description: Event type or prefix.
            example: instance.
        - in: query
          name: since
          description: Event id lower bound for replay.
          schema:
            type: string
            description: Event id lower bound.
            example: 018f3a40-0000-7000-8000-000000000000
      responses:
        "200":
          description: Event stream or JSON snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListEnvelope"
              example:
                data:
                  - id: 018f3a41-0000-7000-8000-000000000000
                    type: instance.started
                    time: "2026-05-05T10:00:00Z"
                    source: incus
                    subject: vm-web-1
                    data: {}
            text/event-stream:
              schema:
                type: string
                description: SSE frames containing Helling event JSON payloads.
              example: "event: instance.started\nid: 018f3a41-0000-7000-8000-000000000000\ndata: {\"type\":\"instance.started\"}\n\n"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: helling_session
      description: Session cookie issued by `POST /auth/login`. Sent automatically by the browser; CLI stores it under `~/.config/helling/session`.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Event:
      type: object
      description: Helling event payload shared by SSE and webhooks.
      required: [id, type, time, source, subject, data]
      properties:
        id:
          type: string
          description: Event UUID.
          format: uuid
          example: 018f3a41-0000-7000-8000-000000000000
        type:
          type: string
          description: Helling event type.
          example: instance.started
        time:
          type: string
          description: Event timestamp.
          format: date-time
          example: "2026-05-05T10:00:00Z"
        source:
          type: string
          description: Source subsystem.
          enum: [incus, podman, helling]
          example: incus
        subject:
          type: string
          description: Resource affected by the event.
          example: vm-web-1
        data:
          type: object
          description: Event-specific metadata.
          additionalProperties: true
          example: {}
...
    EventListEnvelope:
      type: object
      description: Envelope containing events.
      required: [data]
      properties:
        data:
          type: array
          description: Event payload list.
          example:
            - id: 018f3a41-0000-7000-8000-000000000000
              type: instance.started
              time: "2026-05-05T10:00:00Z"
              source: incus
              subject: vm-web-1
              data: {}
          items:
            $ref: "#/components/schemas/Event"
...
    Error:
      type: object
      description: Standard error envelope returned by 4xx and 5xx responses.
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, machine-readable error code.
          example: not_found
        message:
          type: string
          description: Human-readable message.
          example: instance not found
        details:
          type: object
          description: Optional structured context.
          additionalProperties: true
          example:
            field: name